+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     `set_autofire_cps` and `get_autofire_cps`:
   - `Be able to enable/disable incoming Bluetooth connections with:
     `set_bluetooth_enabled`
+- Wii: Add support for Wii Balance Board
 
 ### Changed
 - Console: `set_mouse_scale(value)` where a higher value means faster movement.